Healpix maps with nside > 8192 (#120)
This PR ensures that maps are consistently indexed using 64-bit integers everywhere, to avoid truncation e.g. when indexing maps with high nside (>8192).  The healpix C bindings are moved out of the public API, and the header is modified to use the `ssize_t` integer type for all 64-bit functions.

Fixes #119

+++ maps/include/maps/FlatSkyMap.h
@@ -125,12 +125,12 @@ public:
 
 	std::vector<double> PixelToAngleGrad(size_t pixel, double h=0.001) const;
 
-	void GetRebinAngles(long pixel, size_t scale,
+	void GetRebinAngles(size_t pixel, size_t scale,
 	    std::vector<double> & alphas, std::vector<double> & deltas) const override;
 	void GetInterpPixelsWeights(double alpha, double delta,
-	    std::vector<long> & pixels, std::vector<double> & weights) const override;
+	    std::vector<size_t> & pixels, std::vector<double> & weights) const override;
 
-	std::vector<long> QueryDisc(double alpha, double delta,
+	std::vector<size_t> QueryDisc(double alpha, double delta,
 	    double radius) const override;
 
 	G3SkyMapPtr Rebin(size_t scale, bool norm = true) const override;
